+++ R/zzz.R
@@ -1,0 +1,5 @@
+.onLoad <- function(libname, pkgname) {
+  if (requireNamespace("knitr", quietly = TRUE)) {
+    knitr::knit_engines$set(c4 = c4_knitr_engine)
+  }
+}